agents/orchestrator/pyproject.toml
[project]
name = "orchestrator-builder-agent"
version = "0.1.0"
description = "Orchestrator microservice agent"
requires-python = ">=3.10,<3.14"
dependencies = [
    "google-genai>=1.5.0",
    "google-adk==1.22.*",
    "a2a-sdk==0.3.*",
    "fastapi==0.123.*",
    "uvicorn==0.40.0",
    "google-cloud-logging==3.13.0",
    "opentelemetry-exporter-gcp-trace==1.11.0",
]

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
default = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]


